+++ Tests/SentryTests/SentrySpanTests.swift
@@ -20,7 +20,7 @@ class SentrySpanTests: XCTestCase {
         }
         
         func getSut(client: Client) -> Span {
-            let hub = SentryHub(client: client, andScope: nil, andCrashAdapter: TestSentryCrashWrapper())
+            let hub = SentryHub(client: client, andScope: nil, andCrashAdapter: TestSentryCrashAdapter.sharedInstance())
             return hub.startTransaction(name: someTransaction, operation: someOperation)
         }